refactor: add function to refresh finished status
